Labeling the y-axis
We made a lot of sacrifices in the name of screen real estate, but a poor one
was to leave the y-axis unlabeled. Bandwidth is in Kb or KB, and size is in MB.
Who would guess that? Hell, I need to check the code every time I'm asked.

This is a work in progress. Two missing bits...

  * The graph presently covers the labels if it's long.
  * The labels should be right aligned.