Renaming project to 'seth'
As discussed on tor-dev@ I'm scheming to rename this project from arm to
seth...

  https://lists.torproject.org/pipermail/tor-dev/2015-March/008398.html

I'm taking this slow with the plan to wait another week before renaming the
repository, but I can move forward this much to get a feel of working with the
new name. We can revert later this week if we find a better name.