Use 'GETINFO bw-event-cache' to prepopulate bandwidth graph information
In #14128 added a new controller command that provided up to five minutes of
bandwidth information. This... is awesome! Not only does it negate any need for
reading tor's state file (a dirty hack), but it *greatly* improves our user
experience since we don't need to swap to 15-minute interval graphs!