Rewrite GraphPanel's draw() method
Few months ago I was amid a rewrite of our GraphPanel. Last section left was
its draw() method which was a tangled mess. Said mess is now untangled.

I'm still not completely happy with it. For starters it doesn't have decent
thread safety if attributes are modified in the middle of redrawing but
definitely better than where we were at.