Pretend log file entries are from the currnet year
Log files lack the year so we gotta guess (#15607). We hardcoded 2012 so this
would still work for leap years but this is a pretty gross hack. Better to
pretend the log entries are recent, and if we hit a leap year edge case then
meh. User just won't get log prepopulation.