Drop features.log.showDateDividers
Drop the ability to opt out of grouping by days. Never known this to be useful,
and probably not worth accounting this in our new draw() function.

Also fixing features.log.maxLinesPerEntry and features.logFile. They were
documented as being camel case in our nyxrc.sample (like all options), but in
the code it wasn't.