Change days_since() to day_count()
Earlier I changed the behavior of days_since() to work as advertised, but
actually having an absolute count as it previously did is a lot less error
prone. For instance, the change introduced a narrow timing bug with
deduplication.

Reverting to our previous behavior, but with a more accruate name.