Revise popup dialogs to use 'with'
Our popup dialogs acquire a curses lock for its duration. We were doing it via
a init() and finalize() method, which were always used in a try/finally block.
However, this is exactly the use case for python's 'with' keyword so using that
instead.

Not entirely happy with this code (it's still pending a rewrite), but better.