Revise connection panel's run() method
Changing the method to be identical to the header panel's. We had a sleep to be
more responsive during startup but seems as though this is no longer needed
with the rewritten trackers.

This also drops the '-features.connection.refreshRate' option. Another that
honestly I doubt anyone ever even tried.