Move initial descriptor fetch to FingerprintTracker constructor
To bootstrap the FingerprintTracker we need to make a 'GETINFO ns/all' request
for the router status entries. This is a very costly call (on the order of a
second or so), so we need to be wary of when we make it.

As such, better for this to be done by our constructor's caller rather than
lazy loaded when we need it. This isn't in practice changing anything yet but
will make it easier for us to be more careful with this later. Oh, and it makes
the class simpler. :P