Speed connection panel up by a few seconds
The slowest part of initializing the connection panel is issuing the 'GETINFO
ns/all' query to populate the FingerprintTracker. When the user starts nyx then
immediately goes to the connection page this is a noticeable 3-4 second lag.

Doing this asynchronously in the connection panel's thread, at the *end* of
processing the first results. As a result the user doesn't see this lag at all.
Rather, if they immediately go to the connection page the fingerprints and
nicknames will simply be 'UNKNOWN' for a few seconds. Much better. :P