Move FingerprintTracker to tracker util
When picking apart our old tor_tools I plopped the FingerprintTracker into the
conn_entry because that is its sole caller. However, now that I've cleaned it
up it's time to find a more permanent home.

As the name suggests it's similar to the other trackers, so moving it there.
Also renaming it to ConsensusTracker because it's used to provide performant
lookups of any consensus information we need (not just fingerprints).