Move get_relay_address() to ConsensusTracker
Moving this helper method not only tidies things up, but greatly reduces the
number of GETINFO requests we make (we already process all router status
entries, so why query them again?).

Also, we weren't updating the address/port of relays if they changed in a new
consensus. This fixes that.