Start tracker daemons when they're requested
Why would we want a daemon that isn't running? Presently application resolution
is a no-op because we're talking with a dorment tracker. Lets always give
running instances.

We were also failing to stop one of the trackers, and this tweaks our check so
we skip trackers that were never requested (and hence don't exist).