Move connection type to be tracked by entity
Our get_type() method is a method on individual lines rather than the parent
entity, which is weird. All lines always share the same type, and we almost
always fetch it from entites via...

  my_entity.get_lines()[0].get_type()

Moving it to be tracked on the entity instead, and lines just report whatever
its parent has.