Drop connection panel lock
Not as sure of this as I'd like but moving toward the right design. Rest of our
panels are now generally lock-free and we can do so here as well. Most
important was our entries and that bit is now safe. Small risk around other
self reference usage (mostly by handle_key()) so will need to give this more
thought.