Move field attributes to configuration
As with elsewhere moving our color mappings to our configuration to get it out
of the code.

Also changing the Fields enum to use labels for their values (as we ususally do
with enumerations). This was probably separate so we could use the shorter name
in the nyxrc, but those can use the label names.