Use methods for ConfigEntry attributes rather than a single get()
To get all attributes we had a single ConfigEntry get() method that took an
enum. This was done to avoid needing a separate sort_value() method. Fine idea,
but clunky.

Our connection panel has a sort_value(), that's fine. Also we'll be moving much
of this to Stem, in which case proper methods jive better.