Cache manual when read from 'man tor'
When available we run 'man tor' to get manual information about the user's
specific version. This is a neat trick, but it takes over twice as long as
reading our pre-parsed manual information.

After parsing a manual caching the result to disk so running nyx in the future
is a lot quicker.