Fix curses flickering
Commit 6f91af2 fixed an issue where we froze when viewing the connections page
after first starting, but at the expense of making our whole interface flicker.
This is most noticeable when pressing 'm' to browse the menu.

Can't say I'm entirely clear what's up. The connection panel had its own thread
outside the draw loop so I wouldn't expect it to cause this sort of disruption,
though due to the GIL that's of course possible. Oh well - doing this in the
tracker is just better all around...

  * no more curses flickering
  * connection panel can be viewed right away
  * code's more elegant