Drop 'title_visible' methods from the Panel
I never liked these. They're a hack needed by a couple popups to hide the title
of the panel below them. I couldn't think of an elegant method for doing this,
but did come up with a more localized hack. We now make an extra subwindow that
covers the panel title and erase that.