Merge LabelPanel into the header
The LabelPanel was a one-line panel that usually shows...

  page 1 / 4 - m: menu, p: pause, h: page help, q: quit

... and provides custom messages or prompts for user input. Functionality for
this was spread across the LabelPanel, Controller, Panel, and popups module.
Needless to say this can be simplified.

This has always been part of the header so merging a nicer version of it there.
This is a pretty big overhaul with some unintended consequences - some good,
some bad...

  * Good ones are that our interface now continues to update while awaiting
    keypresses. For example, our graph continues to update when awaiting
    confirmation of something.

  * Step back is that we don't quickly redraw overlapped components as we
    previously did, but this is pretty easily fixable.