Stacktrace when we get a NEWCONSENSUS event
This largely reverts commit df0fc05, but *doesn't* add it back to the
connection panel. That is to say curses still works but the connection panel
doesn't.

Doing this since df0fc05 was deceptively broken (task() was never called since
we weren't a Daemon subclass) *and* getting a NEWCONSENSUS event caused
stacktraces...

  Traceback (most recent call last):
    File "/usr/lib/python2.7/threading.py", line 551, in __bootstrap_inner
      self.run()
    File "/usr/lib/python2.7/threading.py", line 504, in run
      self.__target(*self.__args, **self.__kwargs)
    File "/home/atagar/Desktop/nyx/stem/control.py", line 921, in _event_loop
      self._handle_event(event_message)
    File "/home/atagar/Desktop/nyx/stem/control.py", line 3624, in _handle_event
      listener(event_message)
    File "/home/atagar/Desktop/nyx/nyx/tracker.py", line 815, in _new_consensus_event
      self.update(event.desc)
  AttributeError: 'ConsensusTracker' object has no attribute 'update'