Rewrite show_help_popup()
Bit of an experiment in splitting our Panel class from its responsibilities as
a curses subwindow. With this new pattern our draw function gets a subwindow
argument it can draw into. It only has draw functions (and the curses lock)
within that context.

Thus far results are promissing but don't be surprised if this changes again. :P