Drop _draw_bandwidth_stats()
I can see why we had this but... it's confusing. On small screen resolutions we
replaced our y-axis with the average and total bandwidth usage. This only
applied when rendering bandwidths and it's debatable if this is more valuable
than the y-axis.

On my netbook (which is just 100 colums wide) we still show the averages (just
dropping the total) which imho works just fine. We might try something else if
this makes anyone sad.