Merge msg.misc.event_types into help message
We kept this separate because the old event selection dialog used it. Now that
is no longer the case we can merge this into the message. Also dropping the
'defaults to ...' part since enumerating all those is an eye sore. This value
has a clear enough default since it's shown right away when they start up.

I'm tempted to remove the events from the help message since they're dynamic
and the list we show won't be accurate with new tor releases, but keeping for
the moment.