Lighter routine redraws
Our old draw() method had a 'force' flag that, if false, would cause us to only
be redrawn if our dimensions had changed. We did a 'soft redraw' on every
keypress so dropping this flag was very noticeable, causing flickering and the
interface to be laggy on things like the config panel.

Inverting this flag so we force by default. This is because just about every
time we called redraw we wanted to 'force' it. Soft redraws are the exception,
not the rule.