Split up and expand _handle_key() tests
Good tests but we should split them to be for individula actions when
reasonable. Also improving them a bit. Picking the (40, 80) dimensions was a
poor choice since it makes testing cursor movement difficult.

I suspect Sambuddha thought that was the window dimention. It isn't - it's the
cursor position so actually (0, x) is easier to work with.