Rewrite backlog handling
Mutable function globals are code stink. They break thread safety and are just
about never the right thing to do. The arm 1.4.5 HistoryValidator class was a
lot closer to what we want...

  https://gitweb.torproject.org/nyx.git/tree/src/util/textInput.py?h=release#n103

Cleaning that up and replacing our _handle_history_key() with it. This is
invisible to callers (they're both private helpers). Improving our tests along
the way too.