Occasional stacktraces when starting up
When starting up we occasionally flashed a stacktrace such as...

  File "/home/atagar/Desktop/nyx/nyx/panel/graph.py", line 596, in _update_accounting
    if not nyx_controller.is_paused():
  AttributeError 'NoneType' object has no attribute 'is_paused'

The issue is that our Controller's constructor creates panels which in turn
have threads that use the still-incomplete controller. As such breaking the
constructor into two parts. The first creates a blank instance and the second
adds the panels.

Don't really like this but fixes the issue. I'll revisit this when it's the
controller's turn to be rewitten.