Interpreter panel broke after 100 inputs
Guessing this is a leftover debugging statement?

  raise Exception(self._backlog)

Can't think of any reason for it to be there, and it quite simply broke us when
the user entered their 100th input.

Inverting the order our backlog handler accepts to make our panel a little more
efficient (there's no reason to invert our inputs on every call).