Notify all panels of pausing/unpausing
Initially I thought a listener pattern would make more sense, but just about
all the panels need to be notified when we're paused or unpaused. As such
adding a notification hook to the panel.

This lets us drop get_pause_time() from the Interface since the two panels that
care about this can now track it for themselves.