Fix slider math
Damn this scrollbar method has given me a lot of headaches. Correcting an
off-by-one error that caused the tests to fail on my netbook but pass on my
pc.

I really hope this is the last time I need to weed one of these out. For my
personal reference here's the numbers when working through the test...

  top = 15
  top_index = 21
  size = 30
  self.height = 25

  scrollbar_height = self.height - top - 1 = 25 - 15 - 1 = 9

  bottom_index = top_index + scrollbar_height = 21 + 9 = 30

  slider_top = (scrollbar_height * top_index) // size = (9 * 21) // 30 = 6
  slider_size = (scrollbar_height * (bottom_index - top_index + 1)) // size = (9 * (30 - 21 + 1)) // 30 = 3