W3C validation issues
Addressing issues cited by the W3C validator. Interesting that we don't need
slashes in quite a number of tags - always thought that was best practice...

Only issue left is that <noscript> isn't technically allowed in the head, but
this seems to be pretty well supported and is explicitly allowed in HTML5.