Remove Sphinx site
Now that we're done with the new site dropping the Sphinx copy I was previously
running with. That site was based on stem.torproject.org. Sphinx was a good fit
for Stem since as a library API information is a foremost need, but that isn't
the case for Nyx. Making our own site is both simpler *and* gives us greater
control to make something that looks great.