Svn copy of our logo
Actually, our git repo had a copy of this at our last tagged release but seemed
I deleted it while dropping extranious resources. Good file to have though.

Nyx's logo is from the utilities-system-monitor icon in Oxygen, which is part
of KDE. Not sure where its site is nowadays, but the artwork was under Creative
Commons v3 (A, SA)...

  https://www.atagar.com/freeMedia/