Run static checks in the background
Stem got quite a substantial speedup in its tests by running long-running tasks
in the background. Nyx is a much smaller test suite so not much to gain, but
might as well improve things a tad here too.

Running pyflakes and pycodestyle in the background. On my laptop this drops the
test runtime from 3.8s to 2.4s (37% faster).