Don't clear screen in main draw loop
To avoid screen hiccups from persisting we cleared the screen on every
redraw...

  https://trac.torproject.org/projects/tor/ticket/2830

This seemed fine on my laptop since I wasn't running a real relay, but under
actual load it causes nyx to flicker annoyingly every few seconds. This was a
hack in case of anomalies. Cure was worse than the disease.

We might need to revisit this (... again), but for now this makes the
experience of running nyx on my relay much nicer.