Fetch address at a fixed rate
Stem transparently caches immutable information but our address is a rather
strange critter. It can change but hardly ever does, and is used quite a bit.

Nyx forced Stem to treat this as being cachable but when our address is
unavailable this still accounted for numerous requests. Instead fetching
the our address at a fixed rate of once a minute.

This provides two benefits...

* Avoid pointless requests while there's no network available.
* This way we'll evenually update our address if it changes while running.