Manually handle caching of connections
Connections shouldn't use a lru_cache. If the cache size is less than the
number of connections we have at any point then we consume extra cpu, but if
it's too high we consume more and more memory unnecessarily until it's filled.

Instead evicting from the cache when a connection hasn't been referenced in a
while.