Website menu with less common pages
When looking at our site I just realized it's hard to figure out how to report
bugs. It's in the FAQ, but who's gonna read that?

Adding a 'more' dropdown menu with useful, but less common sections folks will
want (changelog, bug tracker, and contact info for now). The contact page
doesn't exist yet - that's next.

God the styling on this was a pita. I'm really not sure why this negative
margin is necessary but without it the rest of the nav menu gets pushed down.
Stared for hours at this hoping to get something simpler but time to raise the
flag and commit what works.

I really hate the lack of a shared header, but can't say that alone is worth
something like Hyde. :/