Let nyx shut down faster
Especially when first starting up nyx takes a couple seconds to shut down on my
relay. This is because the update method of our connection panel can take a
while if there's many relays. Aborting in the middle of its longest part if the
thread's terminated. This drops our shutdown time in practice to a few
milliseconds.