Consolidate redraws if unable to draw for a while
When the menu's open for a while we block the graph panel from redrawing
itself, causing events to accumilate. Then when the menu's closed we perform
a flurry of updates which makes the graph appear to catch up in fast forward.

Instead, if a panel is unable to draw itself after a second it gives up the
redraw attempt. This way when the menu's closed we perform just a single redraw
that catches us up.

Caught thanks to Guinness on tor-relays@.