Switch installation to setuptools
Got additional testers on #tor-dev and turns out our custom installer causes it
to fail for just about everybody. Dropping the ability to specify a man or
samplerc path to get us back on the road best traveled. Also swapping from
distutil to setuptools so we have entry_points (which we need for a bin
script).