Header stats missing with python3
Ick! took me a while to find this. Python3 changed division from providing ints
to floats. This in turn caused our header panel to call addstr with float x/y
coordinates, which made that funciton fail silently. End result is that when
the header panel showed two columns of stats the right column was missing.

Caught thanks to Stijn.