Notify if nyx cache isn't writable
A couple of folks have had Nyx crash due to their cache being unwritable...

  https://trac.torproject.org/projects/tor/ticket/27938

I suspect this arises when running nyx multiple times as different users that
share the same home directory. Rather than crashing, providing a notice and
falling back to an in-memory cache. Nyx's startup will be a tad slower when
this happens but for most users it probably won't even be noticeable.