Renaming our process could potentially crash nyx
Stem's set_process_name() raises IOErrors when unable to change the name of our
process. In practice this has been moot because it only arises when lengthening
our process name, and nyx always shortens it. But in one instance seems that
wasn't the case...

  https://trac.torproject.org/projects/tor/ticket/27556

Not entirely clear why, but regardless we should definitely catch this if it
arises.