Issue a RESETCONF when a blank value is entered
We lacked a mechanism for resetting values to their initial value. For
instance, take the following scenario...

  * Our ExcludeNodes torrc is unset. The config panel shows it as '<none>'.
  * We change the value to 'stuff'.
  * We want to change it back to '<none>'.

There wasn't a way of doing so. If we cleared the value and hit enter we'd
change it to an empty string. I can't think of any scenarios where users would
*want* to set things to blank strings (and in those edge cases they can always
do so via the torrc).

Now if a user deletes a config value and hits enter we'll reset it, in the
above scenario changing the config option back to '<none>'.