Avoid double counting connections
When a guard, bridge, or exit we scrub sensitive connection data but also
aggregate it into graphs that present the client locales we're being used
from or exit ports.

In reviewing this for an old ticket [1] realized we're counting each
connection made with us rather than unique clients or exit usage. As
such, limiting our counts to one per a remote address.

[1] https://trac.torproject.org/projects/tor/ticket/4281