Use controller's get_start_time() method
Tor just added a getinfo option to retrieve our uptime, so taking advantage of
it here. As a new method naturally this is only present at the moment in the
stem git repo.

From a user perspective this should be transparent. Nyx will now present uptime
when connected to a remote tor process and under the covers we make one fewer
system calls, but that's about all.